MyAgentSystem/requirements.txt
hydra-core
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
langgraph
pydantic
pandas
huggingface-hub
gigachat
python-dotenv
chromadb
sentence-transformers
fastapi
uvicorn
pypdf
beautifulsoup4
markdown
pyngrok
requests
sqlalchemy
alembic
passlib[bcrypt]
python-multipart